requirements.txt
aiofiles==24.1.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.5.2
apimatic-core==0.2.15
apimatic-core-interfaces==0.1.5
apimatic-requests-client-adapter==0.1.6
APScheduler==3.11.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.13.3
billiard==4.2.1
CacheControl==0.12.14
cachetools==5.5.2
celery==5.5.0
certifi==2024.8.30
cffi==1.17.1
channels==4.3.0
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.40.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.3
daphne==4.2.1
deep-translator==1.11.4
deprecation==2.1.0
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.2
django-cloudinary-storage==0.3.0
django-countries==7.6.1
django-debug-toolbar==4.4.6
django-environ==0.10.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
exceptiongroup==1.2.2
faiss-cpu==1.10.0
git-filter-repo==2.47.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
gTTS==2.5.4
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jiter==0.7.0
jsonpickle==3.0.4
jsonpointer==2.4
kombu==5.5.2
langdetect==1.0.9
lxml==5.4.0
msal==1.31.1
msgpack==1.1.0
numpy==2.2.4
oauthlib==3.2.2
openai==1.51.0
openpyxl==3.1.5
packaging==24.1
pandas==2.3.0
pillow==10.4.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==6.30.2
# psycopg2==2.9.7  <-- removed to avoid duplicate driver
psycopg2-binary==2.9.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMuPDF==1.25.5
pyOpenSSL==25.1.0
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.8.2
python-docx==1.2.0
python-dotenv==1.0.1
pytz==2025.2
qrcode==8.2
redis==5.2.1
reportlab==4.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
squareup==37.1.1.20240717
tqdm==4.66.6
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
whitenoise==6.7.0
zope.interface==7.2
pip-tools==7.4.1 
pyproject-hooks==1.2.0

# Billing and Payments
stripe==12.5.1

# AI and Vector Search
openai==1.51.0
pydantic==2.9.2
pydantic-core==2.23.4

# Authentication
django-allauth==0.57.0

# Email Services
postmark==0.0.1
boto3==1.35.0  # For AWS SES

# Vector Database (PostgreSQL with pgvector)
psycopg2-binary==2.9.7
dj-database-url==2.3.0

# Additional utilities
jinja2==3.1.4  # For email templates
celery==5.5.0  # For background tasks
redis==5.2.1  # For Celery broker